FG-03-08: Refactor hasValue
