+++ src/views/LettersIndex.vue
@@ -136,6 +136,7 @@
 <script>
 import { mapActions } from "vuex";
 import { dataService } from "../shared";
+import tableService from "@/services/table-service";
 import SelectAutoComplete from "../components/SelectAutoComplete.vue";
 import SelectYears from "../components/SelectYears.vue";
 import {
@@ -408,42 +409,20 @@ export default {
       return idNameMap;
     },
 
-    hasValue(item, property, value) {
-      // Check if a property (Array or String) contains or is equal to a value
-      const stack = property.split(".");
-      var prop = item.properties;
-      while (stack.length > 1) {
-        prop = prop[stack.shift()];
-      }
-      prop = prop[stack.shift()];
-
-      if (prop instanceof Array) {
-        return prop.includes(value);
-      }
-      return prop === value;
-    },
-
-    filterItems(objectArray, property, value) {
-      const filtered = objectArray.filter((item) =>
-        this.hasValue(item, property, value)
-      );
-      return filtered;
-    },
-
     filterLetters(rows, terms) {
       if (terms.recipient !== "") {
         rows = rows.filter((r) =>
-          this.hasValue(r, "recipient", terms.recipient)
+          tableService.hasValue(r, "recipient", terms.recipient)
         );
       }
       if (terms.placeSent !== "") {
         rows = rows.filter((r) =>
-          this.hasValue(r, "place.sent", terms.placeSent)
+          tableService.hasValue(r, "place.sent", terms.placeSent)
         );
       }
       if (terms.placeReceived !== "") {
         rows = rows.filter((r) =>
-          this.hasValue(r, "place.received", terms.placeReceived)
+          tableService.hasValue(r, "place.received", terms.placeReceived)
         );
       }
       if (terms.years.length > 0) {
