+++ tests/unit/fixtures/letter-index-item.js
@@ -1,0 +1,22 @@
+export const letterIndexItem = {
+  entity: "letters",
+  id: "G000093",
+  properties: {
+    comments: ["nczb_dlm_ylb", "nalr_3nm_ylb", "nq51_kpm_ylb"],
+    date: "1860-01-11",
+    incipit: "Ihr Brief vom 24. December, mein hochverehrter und teurer Herr,",
+    recipient: ["G001644"],
+    mentioned: {
+      persons: ["G067651", "G001643", "G001645"],
+      places: ["G000646", "ed_hqc_lrv_gmb", "G000860"],
+      works: ["G004289", "G003476"],
+    },
+    place: {
+      received: "G000636",
+      sent: "G000763",
+    },
+    resp: ["Steinsiek", "Costea"],
+    sender: ["G001200"],
+    title: "Ferdinand Gregorovius an Hermann von Thile in Berlin. Rom, 11. Januar 1860"
+  }
+};
